perf(test): split model_runner_v2 basic e2e into dense and spec files
Move spec-decode cases to test_spec_decode.py so CI can schedule them separately. Cap max_num_batched_tokens at 256 to shrink torch.compile range.
