+++ tests/e2e/pull_request/one_card/model_runner_v2/test_basic.py
@@ -20,40 +20,26 @@ from unittest.mock import patch
 
 import pytest
 from vllm import SamplingParams
-from vllm.v1.metrics.reader import Counter, Vector
 
 from tests.e2e.conftest import VllmRunner, wait_until_npu_memory_free
-from tests.e2e.pull_request.one_card.model_runner_v2.utils import calculate_acceptance_per_pos
+from tests.e2e.pull_request.one_card.model_runner_v2.utils import (
+    DEFAULT_PIECEWISE,
+    FULL_DECODE_ONLY,
+    MAX_NUM_BATCHED_TOKENS,
+    MAX_NUM_SEQS,
+    PROMPTS,
+)
 
 DENSE_EAGER_MODELS = ["Qwen/Qwen3-0.6B"]
-DEEPSEEK_W8A8 = "vllm-ascend/DeepSeek-V2-Lite-W8A8"
+DEEPSEEK_W8A8_MODEL = "vllm-ascend/DeepSeek-V2-Lite-W8A8"
 
-MAIN_MODELS = ["LLM-Research/Meta-Llama-3.1-8B-Instruct"]
-EGALE_MODELS = ["vllm-ascend/EAGLE-LLaMA3.1-Instruct-8B"]
-DFLASH_MAIN_MODEL = ["Qwen/Qwen3-8B"]
-DFLASH_MODELS = ["z-lab/Qwen3-8B-DFlash-b16"]
-DSPARK_MAIN_MODEL = ["Qwen/Qwen3-8B"]
-DSPARK_MODELS = ["deepseek-ai/dspark_qwen3_8b_block7"]
-MTP_MODELS = ["wemaster/deepseek_mtp_main_random_bf16"]
-
-# Keep default cudagraph_mode (FULL_AND_PIECEWISE when unspecified) but pin
-# capture sizes. The workload is 4 prompts; leaving sizes unset enumerates
-# graphs up to min(max_num_seqs * (1+K), 512) and dominates runtime.
-CUDAGRAPH_CAPTURE_SIZES = [4, 8]
-FULL_DECODE_ONLY = {
-    "cudagraph_mode": "FULL_DECODE_ONLY",
-    "cudagraph_capture_sizes": CUDAGRAPH_CAPTURE_SIZES,
-}
-DEFAULT_PIECEWISE = {"cudagraph_capture_sizes": CUDAGRAPH_CAPTURE_SIZES}
-# Matches the 4-prompt batch; also shrinks KV/draft padding vs default 256.
-MAX_NUM_SEQS = 8
-
-# One graph mode per spec method; both modes still covered on Qwen3-0.6B.
-# DeepSeek-V2-Lite is a single graph-mode smoke (no output golden).
+# One graph mode per spec method lives in test_spec_decode.py; both modes
+# are still covered here on Qwen3-0.6B. DeepSeek-V2-Lite is a single
+# graph-mode smoke (no output golden).
 DENSE_GRAPH_CASES = [
     pytest.param("Qwen/Qwen3-0.6B", FULL_DECODE_ONLY, id="qwen3-full_decode_only"),
     pytest.param("Qwen/Qwen3-0.6B", DEFAULT_PIECEWISE, id="qwen3-default_full_and_piecewise"),
-    pytest.param(DEEPSEEK_W8A8, FULL_DECODE_ONLY, id="dsv2-full_decode_only"),
+    pytest.param(DEEPSEEK_W8A8_MODEL, FULL_DECODE_ONLY, id="dsv2-full_decode_only"),
 ]
 
 
@@ -66,13 +52,6 @@ def test_qwen3_dense_eager_mode(
     max_tokens: int,
     enforce_eager: bool,
 ) -> None:
-    prompts = [
-        "Hello, my name is",
-        "The president of the United States is",
-        "The capital of France is",
-        "The future of AI is",
-    ]
-
     sampling_params = SamplingParams(
         max_tokens=max_tokens,
         temperature=0.5,
@@ -89,194 +68,11 @@ def test_qwen3_dense_eager_mode(
         model,
         max_model_len=1024,
         max_num_seqs=MAX_NUM_SEQS,
+        max_num_batched_tokens=MAX_NUM_BATCHED_TOKENS,
         enforce_eager=enforce_eager,
         async_scheduling=True,
     ) as runner:
-        runner.generate(prompts, sampling_params)
-
-
-@pytest.mark.parametrize("model", MAIN_MODELS)
-@pytest.mark.parametrize("eagle_model", EGALE_MODELS)
-@pytest.mark.parametrize("max_tokens", [32])
-@pytest.mark.parametrize("enforce_eager", [False])
-@patch.dict(os.environ, {"VLLM_USE_V2_MODEL_RUNNER": "1"})
-def test_egale_spec_decoding(
-    model: str,
-    eagle_model: str,
-    max_tokens: int,
-    enforce_eager: bool,
-) -> None:
-    prompts = [
-        "Hello, my name is",
-        "The president of the United States is",
-        "The capital of France is",
-        "The future of AI is",
-    ]
-    num_speculative_tokens = 3
-    sampling_params = SamplingParams(max_tokens=max_tokens, temperature=0.0)
-    with VllmRunner(
-        model,
-        max_model_len=1024,
-        max_num_seqs=MAX_NUM_SEQS,
-        enforce_eager=enforce_eager,
-        disable_log_stats=False,
-        async_scheduling=True,
-        speculative_config={
-            "model": eagle_model,
-            "method": "eagle",
-            "num_speculative_tokens": num_speculative_tokens,
-        },
-        compilation_config=FULL_DECODE_ONLY,
-    ) as runner:
-        runner.model.generate(prompts, sampling_params)
-        metrics = runner.model.get_metrics()
-
-    acceptance_per_pos = calculate_acceptance_per_pos(
-        metrics,
-        num_speculative_tokens,
-        Counter,
-        Vector,
-    )
-    golden = [0.43, 0.13, 0.05]
-    match = all(abs(a - b) < 0.1 for a, b in zip(acceptance_per_pos, golden))
-    assert match, f"acceptance_per_pos {acceptance_per_pos} does not match golden {golden}"
-
-
-@pytest.mark.parametrize("model", DFLASH_MAIN_MODEL)
-@pytest.mark.parametrize("dflash_model", DFLASH_MODELS)
-@pytest.mark.parametrize("max_tokens", [32])
-@pytest.mark.parametrize("enforce_eager", [False])
-@patch.dict(os.environ, {"VLLM_USE_V2_MODEL_RUNNER": "1"})
-def test_dflash_spec_decoding(
-    model: str,
-    dflash_model: str,
-    max_tokens: int,
-    enforce_eager: bool,
-) -> None:
-    prompts = [
-        "Hello, my name is",
-        "The president of the United States is",
-        "The capital of France is",
-        "The future of AI is",
-    ]
-
-    num_speculative_tokens = 7
-    sampling_params = SamplingParams(max_tokens=max_tokens, temperature=0.0)
-    with VllmRunner(
-        model,
-        max_model_len=1024,
-        max_num_seqs=MAX_NUM_SEQS,
-        enforce_eager=enforce_eager,
-        disable_log_stats=False,
-        async_scheduling=True,
-        speculative_config={
-            "model": dflash_model,
-            "method": "dflash",
-            "num_speculative_tokens": num_speculative_tokens,
-        },
-        compilation_config=DEFAULT_PIECEWISE,
-    ) as runner:
-        runner.model.generate(prompts, sampling_params)
-        metrics = runner.model.get_metrics()
-
-    acceptance_per_pos = calculate_acceptance_per_pos(
-        metrics,
-        num_speculative_tokens,
-        Counter,
-        Vector,
-    )
-
-    golden = [0.51, 0.16, 0.07, 0.07, 0.01, 0.01, 0.0]
-    match = all(abs(a - b) < 0.1 for a, b in zip(acceptance_per_pos, golden))
-    assert match, f"acceptance_per_pos {acceptance_per_pos} does not match golden {golden}"
-
-
-@pytest.mark.parametrize("model", DSPARK_MAIN_MODEL)
-@pytest.mark.parametrize("dspark_model", DSPARK_MODELS)
-@pytest.mark.parametrize("max_tokens", [32])
-@pytest.mark.parametrize("enforce_eager", [False])
-@patch.dict(os.environ, {"VLLM_USE_V2_MODEL_RUNNER": "1"})
-@wait_until_npu_memory_free(target_free_percentage=0.8)
-def test_dspark_spec_decoding(
-    model: str,
-    dspark_model: str,
-    max_tokens: int,
-    enforce_eager: bool,
-) -> None:
-    prompts = [
-        "Hello, my name is",
-        "The president of the United States is",
-        "The capital of France is",
-        "The future of AI is",
-    ]
-
-    num_speculative_tokens = 7
-    sampling_params = SamplingParams(max_tokens=max_tokens, temperature=0.0)
-    with VllmRunner(
-        model,
-        max_model_len=1024,
-        max_num_seqs=MAX_NUM_SEQS,
-        enforce_eager=enforce_eager,
-        disable_log_stats=False,
-        async_scheduling=True,
-        speculative_config={
-            "model": dspark_model,
-            "method": "dspark",
-            "num_speculative_tokens": num_speculative_tokens,
-        },
-        compilation_config=FULL_DECODE_ONLY,
-    ) as runner:
-        runner.model.generate(prompts, sampling_params)
-        metrics = runner.model.get_metrics()
-
-    acceptance_per_pos = calculate_acceptance_per_pos(
-        metrics,
-        num_speculative_tokens,
-        Counter,
-        Vector,
-    )
-    golden = [0.84, 0.48, 0.32, 0.20, 0.09, 0.09, 0.02]
-    match = all(abs(a - b) < 0.1 for a, b in zip(acceptance_per_pos, golden))
-    assert match, f"acceptance_per_pos {acceptance_per_pos} does not match golden {golden}"
-
-
-@pytest.mark.parametrize("model", MTP_MODELS)
-@pytest.mark.parametrize("max_tokens", [32])
-@pytest.mark.parametrize("enforce_eager", [False])
-@patch.dict(os.environ, {"VLLM_USE_V2_MODEL_RUNNER": "1"})
-@wait_until_npu_memory_free(target_free_percentage=0.8)
-def test_mtp_spec_decoding(
-    model: str,
-    max_tokens: int,
-    enforce_eager: bool,
-) -> None:
-    # The MTP draft head has random weights, so acceptance is ~0 and there is
-    # no trained golden to compare against -- this is a smoke test (assert only
-    # that the MTP MLA propose->verify loop produces output).
-    prompts = [
-        "Hello, my name is",
-        "The president of the United States is",
-        "The capital of France is",
-        "The future of AI is",
-    ]
-    num_speculative_tokens = 3
-    sampling_params = SamplingParams(max_tokens=max_tokens, temperature=0.0)
-    with VllmRunner(
-        model,
-        max_model_len=1024,
-        max_num_seqs=MAX_NUM_SEQS,
-        enforce_eager=enforce_eager,
-        async_scheduling=True,
-        enable_expert_parallel=True,
-        speculative_config={
-            "method": "mtp",
-            "num_speculative_tokens": num_speculative_tokens,
-        },
-        compilation_config=DEFAULT_PIECEWISE,
-    ) as runner:
-        outputs = runner.model.generate(prompts, sampling_params)
-
-    assert len(outputs) == len(prompts)
+        runner.generate(PROMPTS, sampling_params)
 
 
 @pytest.mark.parametrize("model, compilation_config", DENSE_GRAPH_CASES)
@@ -290,22 +86,16 @@ def test_qwen3_dense_graph_mode(
     enforce_eager: bool,
     compilation_config: dict,
 ) -> None:
-    prompts = [
-        "Hello, my name is",
-        "The president of the United States is",
-        "The capital of France is",
-        "The future of AI is",
-    ]
-
     sampling_params = SamplingParams(max_tokens=max_tokens, temperature=0.0)
     with VllmRunner(
         model,
         max_model_len=1024,
         max_num_seqs=MAX_NUM_SEQS,
+        max_num_batched_tokens=MAX_NUM_BATCHED_TOKENS,
         enforce_eager=enforce_eager,
         compilation_config=compilation_config,
     ) as runner:
-        outputs = runner.model.generate(prompts, sampling_params)
+        outputs = runner.model.generate(PROMPTS, sampling_params)
 
     if model != "Qwen/Qwen3-0.6B":
         return
